set up router and outlet
