+++ myapp/src/components/ChatList/index.jsx
@@ -1,29 +1,49 @@
 import * as React from 'react';
-import { Link, Outlet } from "react-router-dom"
 import List from '@mui/material/List';
 import ListItem from '@mui/material/ListItem';
 import ListItemText from '@mui/material/ListItemText';
 import ListItemAvatar from '@mui/material/ListItemAvatar';
 import { Avatar } from '@mui/material/';
+import { Link, Outlet } from 'react-router-dom';
 
-export default function ChatList({ data }) {
+let chats = [
+    {
+        id: "1",
+        name: "Antonio"
+    },
+    {
+        id: "2",
+        name: "Tomas"
+    },
+    {
+        id: "3",
+        name: "Angelina"
+    },
+    {
+        id: "4",
+        name: "Brandy"
+    }
+];
+
+export default function ChatList() {
     return (
+        <>
         <List sx={{ width: '100%', maxWidth: 360, bgcolor: 'background.paper' }}>
-            {data.map((chat) => {
+            {chats.map((chat) => {
                 return (
-                    <>
-                        <Link key={chat.id} to={`/chat${chat.id}`}>
-                            <ListItem alignItems="flex-start" sx={{ borderTop: "solid 1px" }}>
-                                <ListItemAvatar>
-                                    <Avatar alt={chat.name} src="/static/images/avatar/1.jpg" />
-                                </ListItemAvatar>
+                        <ListItem key={chat.id} alignItems="flex-start" sx={{ borderTop: "solid 1px" }}>
+                            <ListItemAvatar>
+                                <Avatar alt={chat.name} src="/static/images/avatar/1.jpg" />
+                            </ListItemAvatar>
+                            <Link key={chat.id} to={`/chats/chat${chat.id}`}>
                                 <ListItemText primary={chat.name} />
-                            </ListItem>
-                        </Link>
-                        <Outlet />
-                    </>
+                            </Link>
+                            <button onClick={()=> console.log(chat.id)}>x</button>
+                        </ListItem>
                 )
             })}
         </List>
+        <Outlet />
+        </>
     )
 }
